+++ modules/utils/treeUtils.js
@@ -99,20 +99,24 @@ export const removeIsLockedInTree = (tree) => {
 
 
 /**
- * Set correct `path` in every item
+ * Set correct `path` and `id` in every item
  * @param {Immutable.Map} tree
  * @return {Immutable.Map} tree
  */
 export const fixPathsInTree = (tree) => {
   let newTree = tree;
 
-  function _processNode (item, path, lev) {
+  function _processNode (item, path, lev, nodeId) {
     if (!item) return;
-    const _id = item.get("id");
-    const itemPath = path.push(item.get("id"));
-    const currItemPath = item.get("path");
-    if (!currItemPath || !currItemPath.equals(itemPath)) {
+    const currPath = item.get("path");
+    const currId = item.get("id");
+    const itemId = currId || nodeId;
+    const itemPath = path.push(itemId);
+    if (!currPath || !currPath.equals(itemPath)) {
       newTree = newTree.setIn(expandTreePath(itemPath, "path"), itemPath);
+    }
+    if (!currId) {
+      newTree = newTree.setIn(expandTreePath(itemPath, "id"), itemId);
     }
 
     const children = item.get("children1");
@@ -124,8 +128,8 @@ export const fixPathsInTree = (tree) => {
           new Immutable.OrderedMap(children)
         );
       }
-      children.map((child, _childId) => {
-        _processNode(child, itemPath, lev + 1);
+      children.map((child, childId) => {
+        _processNode(child, itemPath, lev + 1, childId);
       });
     }
   }
@@ -139,15 +143,15 @@ export const fixPathsInTree = (tree) => {
 export const fixEmptyGroupsInTree = (tree) => {
   let newTree = tree;
 
-  function _processNode (item, path, lev) {
+  function _processNode (item, path, lev, nodeId) {
     if (!item) return false;
-    const id = item.get("id");
-    const itemPath = path.push(item.get("id"));
+    const itemId = item.get("id") || nodeId;
+    const itemPath = path.push(itemId);
 
     const children = item.get("children1");
     if (children) {
-      const allChildrenGone = children.map((child, _childId) => {
-        return _processNode(child, itemPath, lev + 1);
+      const allChildrenGone = children.map((child, childId) => {
+        return _processNode(child, itemPath, lev + 1, childId);
       }).reduce((curr, v) => (curr && v), true);
       if ((children.size == 0 || allChildrenGone) && lev > 0) {
         newTree = newTree.deleteIn(expandTreePath(itemPath));