+++ tests/support/inits.js
@@ -2,6 +2,21 @@
 import { Utils } from "react-awesome-query-builder";
 const { uuid } = Utils;
 
+export const tree_with_empty_group = {
+  type: "group",
+  id: uuid(),
+  children1: {
+    [uuid()]: {
+      type: "group",
+      properties: {
+        conjunction: "AND",
+        not: false
+      },
+      children1: {}
+    },
+  }
+};
+
 export const tree_with_number = {
   type: "group",
   id: uuid(),